@@ -16,17 +16,17 @@
   "private": true,
   "dependencies": {
     "@agm/core": "^1.0.0-beta.5",
-    "@angular/animations": "^7.2.0",
+    "@angular/animations": "^8.2.2",
     "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.0",
-    "@angular/compiler": "^7.2.0",
-    "@angular/core": "^7.2.0",
+    "@angular/common": "^8.2.2",
+    "@angular/compiler": "^8.2.2",
+    "@angular/core": "^8.2.2",
     "@angular/fire": "^5.1.1",
-    "@angular/forms": "^7.2.0",
+    "@angular/forms": "^8.2.2",
     "@angular/http": "^7.2.0",
-    "@angular/platform-browser": "^7.2.0",
-    "@angular/platform-browser-dynamic": "^7.2.0",
-    "@angular/router": "^7.2.0",
+    "@angular/platform-browser": "^8.2.2",
+    "@angular/platform-browser-dynamic": "^8.2.2",
+    "@angular/router": "^8.2.2",
     "@ngrx/effects": "^7.4.0",
     "@ngrx/router-store": "^7.4.0",
     "@ngrx/store": "^7.4.0",
@@ -38,15 +38,15 @@
     "ng2-semantic-ui": "^0.9.7",
     "ngx-smart-modal": "^7.1.1",
     "ngx-toastr": "^10.0.2",
-    "rxjs": "^6.0.0",
+    "rxjs": "^6.5.2",
     "tslib": "^1.9.0",
-    "zone.js": "~0.9.0"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.13.8",
     "@angular/cli": "~7.3.8",
-    "@angular/compiler-cli": "^7.2.0",
-    "@angular/language-service": "^7.2.0",
+    "@angular/compiler-cli": "^8.2.2",
+    "@angular/language-service": "^8.2.2",
     "@types/jasmine": "~3.3.12",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~11.13.2",
@@ -61,7 +61,7 @@
     "protractor": "~5.4.2",
     "ts-node": "~8.0.3",
     "tslint": "~5.15.0",
-    "typescript": "~3.2.2",
+    "typescript": "~3.5.3",
     "semantic-ui": "^2.4.2"
   }
 }
